\name{coef.corHerit}
\alias{coef.corHerit}
\title{Coefficients of a corHerit Object
}
\description{This method function extracts the coefficients associated with the correlation structure represented by object.
}
\usage{

## S3 method for class 'corHerit'
\method{coef}{corHerit}(object, ...)

}
\arguments{
  \item{object}{an object inheriting from class \code{corHerit}, representing a correlation structure.
}
  \item{\dots}{
some methods for this generic require additional arguments. None are used in this method.
}
}
\value{
a vector of parameter values

}
\references{
Pinheiro, J. C. and Bates, D. M. (2000), Mixed-Effects Models in S and S-PLUS, Springer, New York.}


